app/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
pytz==2024.1
